smv: Add ACP port to convolution, remove cache support.
The previous access mechanism was far too complicated and made no sense
anyways, so it's all been removed. It's currently been replaced by only
an ACP port on the outputs.

The ACP port allows the kernel to commit the finished output pixels
directly over ACP when it is finished, while still using local
scratchpads to store the temporary unfinished partial sums. There is a
slight hack right now, because it actually uses the activation function
to perform the ACP stores, as opposed to a dedicated "acpStore" function
(like dmaStore), but this does let us hide more latency and reduce the
requirements on ACP bandwidth.

Change-Id: If78e69fec3ed286db5b9ba1b628996a7bb228b29