smv: Handle zeropadding in the logic directly.
Rather than having the CPU manually add zeropadding, we detect which
rows and columns are in the padding region. If so, we feed zero to the
activation register. The rest of the computation still proceeds as
usual, so we don't actually save any time on the HW, but we do save CPU
time and involvement.

Change-Id: I2a416c12688db4aea4ce744bf74e4e211eb41888