smv: Add ACP port to inner product.
This works similarly as the convolution kernel. Finished output pixels
can be written directly out to the host via ACP, as can partial sums at
the end of an iteration if using compressed weights.

One possible optimization on top of this is that after decompression,
use ACP to pull in the partial sums from the last iteration into the
relevant scratchpad, instead of using DMA.

Change-Id: I09c538702a9b774fe119dc5337c6345389f612de