smv: Write back output feature maps after every intra-tile iteration.
All the partial sums and finished sums are still initially written to
the internal scratchpads. When this is done, depending out the specified
output requirement, the data is either copied out via DMA or via ACP,
for which new coherentLoad/coherentStore functions have been added.

This is preparation for CPU-accelerator overlap on the activation
function.

Change-Id: I808da9f0f81e51bd2e956eb7a2c3938fee35be15