smv: Rename some function arguments in conv/inner product.
We used to use dma_* only for DMA loads and stores, but now we use them
for ACP coherent loads/stores as well, so it makes more sense to prefix
them with host_. This also means we can get rid of the additional
acp_results parameter, as it is superseded by host_results.

NOTE: inner product does not support using a local cache! If output_req
or input_req == IO_CACHE, it will fallback to using ACP. This is only
because I'm lazy and don't feel like writing out all the possible
combiations...if we need this, we can add it later.

Change-Id: I070c20b698f1ed8af38641146d7e8ee8db87dcea