smv: Implement packed FP16 activation functions.
This implementation unpacks the FP16 data into a temporary FP32 buffer,
runs a 128-bit SIMD implementation of the activation functions when
possible, and then repacks it back into FP16. We only use up to SSE3
instructions, except for the F16C instructions required to
unpack/repack, for gem5 compatibility.

This is meant for the CPU to use, not for Aladdin.

Change-Id: I25fa434c1e3d0f82f57e5a6cadca48d7e067466c