utility: Reorganize fp16_utils.h.
At this point, we have several different environments that need to build
and run with the half precision conversion functions, and they have
different constraints on how the conversion functions. The biggest
problem is running under gem5: we cannot enable the F16C instructions
via the -mf16c flag or that will also implicitly enable the AVX2
instruction and register set, which gem5 does not support. Instead,
we need to use inline assembly to emit the supported conversion iforms.

This commit adds smaug "builtin" functions that mimic the compiler
builtins in their usage for this scenario. They are enabled with the
__USE_F16C_ANYWAYS__ macro. Any scenario that does not have either
__F16C__ or this macro will fallback to the SW emulation provided by the
FP16 library.

This commit also turns off the ABI warnings that we saw in the gem5
binary due to not enabling AVX while using 256-bit vectors.

Change-Id: I93e74ef7fa21de9a1c080e85e08cf58c91211c70