core: Rename the data structures used to store differently typed data.
Weights for different layers can arrive in many different formats -
uncompressed fp32, uncompressed fp16, CSR, or packed CSR. We'll be
applying this data structure to all the activations as well, so we need
a generic name.

This also move the init_data_list functions out of the top level
nnet_fwd.c and into utility.

Change-Id: Ia80d2f1bac9fb1399cdaa491d4f08c58d434e4bc