arch: Generalize SMAUG to accept a variety of data types [2/5].
This applies all the required changes to the composable backend. This
also uncovered the reason why composable was failing the unit tests on
the real data tests: we reused the parameter archive files from
monolithic, but did not set TRANSPOSE_WEIGHTS=1 when building composable.
composable is now passing all unit tests again.

Change-Id: I56612e3c6a592d81d9ebb990746d74db91cc33ff