arch: Generalize SMAUG to accept a variety of data types [5/5].
This commit migrates SMV to the new data_list-oriented API.

The data-list API was intended to allow SMV to use different data types
for activations and weights, but as a first pass, this commit continues
to use FP32 as the only data type, since adding FP16 will require many
more changes. The next series of commits will bring FP16 support to SMV
to take advantage of the new data list structure.

Change-Id: I04a66edd2e593f1ac6879ec6f6e2ca651669c7c7