core: Rename uarray_t to fp16array_t.
uarray_t was only ever meant to store packed fp16 values (we have
iarray_t for integral values).

This commit also changes the meaning of the len parameter in
init_fp16array: to keep API semantics consistent with init_farray, it
also refers to the number of __FP16__ elements, not the number of 32-bit
integers that would be required to store them (which now gets calculated
internally).

Change-Id: I32b0d3a81763637021973c96af0245c2095f103d