smv: Support for FP16 activations in convolution.
The HW will unpack the FP16 data as it loads them from the host memory
into the internal UMEM. Weights have already been converted to FP16.
All math is done in FP32, and when the HW is finished, it repacks the
results into FP16 and sends them back to the host. Given our
output-stationary dataflow and the fact that we don't support tiling in
the channel dimension, this does not change the accuracy of the output.

Change-Id: Iefeb9f441f4a102ea958b85699a5179ddadaba9a