smv: Support FP16 activations in inner product.
The overall flow is the same as convolution. However, one difference is
that the FC tiling strategy for FC layers leads to precision loss over
iterations, because we send back FP16 partial sums when one HW iteration
is done, but those partial sums are later reused (after having lost
precision) when adding the bias and running the activation function.

Change-Id: I8c00599c3d8f66fc9f2b1e27e4a0b3ea5c377dd7