smiv: Fix corner case in coherentLoad/coherentStore.
If the transfer size is not a multiple of 32 bytes (e.g. 48 bytes), then
the last remaining 31 bytes or less will dropped. Assuming a multiple of
32 bytes for the transfer size worked fine before we started introducing
packed FP16 data formats. Now, a 32-byte vector of 8 FP32 is compressed
into a 16-byte vector of FP16, so we can potentially require 16-byte
transfers.

To solve this, we do a read-modify-write operation on the last remaining
bytes. This is fine for loads but slow for stores, since we have to read
the host memory, update the parts we want, and then send it back, but it
is simple and the overall cost is low relative to the rest of the data
transfer.

Change-Id: Ib07868a6bdbe5177d2453f8a2f8bb70d361bd58b