tests: Rewrite FP comparison logic to support FP16.
A few things need to be considered for FP16.

  1) Limited dynamic range (anything above 65K is INF).
  2) Limited integral representation resolution above 2k.
  3) In general, much higher percent error when compared to a value
     generated using 32-bit math.

This commit factors out the comparison logic into a functor class, which
can have different thresholds and behavior for different backends.
This eliminates the global error thresholds we had before. It is also
easier to add new comparison checkers in the future.

Change-Id: I9095e71ab593fff03dad46c57f96e44d67211762