smv: Refactor the sampling handling logic.
All sampling-related metadata is now computed ahead of time and stored
in the tiling configuration structures. These structures have been
augmented with execute bits and upscaling factors. During execution, we
can simply check the execute bit to know whether we should skip a
particular output tile or HW pass, and if not, we set the profiling log
sampling fields with the precomputed upscaling factors.

This has two major benefits:

  1) Changes to how the sampling algorithm works is no longer entangled
     with the actual HW dispatch code, reducing the number of variables
     needed to dynamically track metadata and determine whether to skip
     a loop iteration or not.
  2) We no longer need to fix up the profiling log in post-processing,
     because all the upscaling factors are now attached to the correct
     profiling entries.

As part of this commit, the following two changes to the sampling
algorithm:

  1) In addition to always executing the last output tile, we also also
     always execute the first, because the first HW pass of the first
     output tile may be responsible for loading input activations.
  2) We no longer require running the first *and* last HW passes. In the
     past, the last HW pass was responsible for copying data back to the
     host, but now we do it on every pass.

Change-Id: I6ce069487d44b0fb5b3aee25f76a96fab6a6297c