smv: Support blocked CSR weights for FC layers.
The FC block requires us to store at least VECTOR_SIZE rows of weights
in order to maximize PE utilization.  If the UMEM size is unsufficient
to store VECTOR_SIZE rows of weights, then we decompress the weights and
recompress them in blocks of N columns, where N = SMV_UMEM_SIZE /
VECTOR_SIZE / sizeof(float). The weights are stored in a data list of
length num_blocks + 1 (for the biases).

Change-Id: I63f3b1bc232daec6c0a1523669976d18e5f59deb