smv: Support rowwise and column-wise tiling in inner product.
This is a major refactoring of the SMV inner product handling code. In
addition to strip mining the weights, if even a single strip is too
large to fit in the private scratchpads, we can take advantage of the
blocked weights created in the previous commit and tile the work
column-wise in addition to row-wise. Column-wise tiling configurations
are stored in a tile struct, while row-wise tiling configurations are
stored in a strip struct.

Change-Id: Ica80d4f171d0e56fc72393cba0fbb9ab960efadd