smv: Add a 16-byte coherent load/store function.
The private scratchpads are partitioned at 32-byte granularity, so we
ordinarily cannot do a 16-byte load or store. However, for the smallest
tiling configurations, we sometimes need to store merely 8 FP16 partial
sums, which is a 16-byte quantity. To do so, we either perform a
read-modify-write operation (to store into the scratchpad) or we read
out a 32-byte value, extract the 16-bytes we want, and store that to the
host (which can accept any size transaction).

Change-Id: I77245f4b66d438b730c3bd12daf5dedfa422c371