smv: Add an element-wise operation block.
This block is used to add the bias and run activation functions
separately from the inner product block. In the past, we had always
fused these two operations, but the increasing tiling and variation in
the layer shapes makes this harder and harder to do.

Now we can have the inner product only worry about computing the correct
GEMM, and we can handle the different bias dimensions and DMA behavior
separately. This greatly simplifies the logic when running CSR
decompression, since that no longer needs to handle copying the biases
and partial sums over in addition to the weights.

Change-Id: I61728a8ee0fc60c24020f67c1d42a720a3defd0d