smv: Ensure DMA stores sizes are multiples of 32-bytes.
This was not required back when we would always round the transfer size
to 64 bytes. However, there are now scenarios where a 16 byte ACP
transfer is required. But we don't have the fine grained control over
DMA that we have for ACP, so we need to ensure that DMA will not try to
access a scratchpad with a data width smaller than its word size.

Change-Id: I299a3e5699141b2f9996111780f5163b232f13cc