smv: Fix some bugs with eltwise_hw and inner product.
In particular, the inner product was sometimes mistakenly not setting
output_req to a value other than IO_NONE, even though it knows
decompression is required, because it would just rely on the curr_layer
descriptor.

This also attaches the "FC" type to the elementwise operation; before
it was an uninitialized value.

Change-Id: Id984435351c9d82a3feda467d851256b8b54ee08