smv: Support input tiling for pooling layers.
This input tiling is a channel-wise tiling strategy, because the pooling
hardware uses a blocked NHWC input data format. Each block is eight
channels deep, so each input tile is at minimum a RxCx8 volume, where R
and C are the full row and col dimensions of the input image. We do not
strip mine at the row level for each input image (although we could
implement that as yet another level of tiling if need be).

This has been manually tested for a large pooling layer and varying UMEM
sizes from 128KB to 2MB.

Change-Id: I32bf407682c83a5b8f1e22d6a83c5d6588e8eed5