smv: Fix array mapping bug in convolution.
We were mapping too much memory for the weights location. weights_loc
will increment up with each output tile, and we only need to map the
region required by the current output tile, but instead we were always
mapping a region with the size of the entire weights. As a result, this
region would eventually overlap with other arrays, resulting in Aladdin
resolving addresses to incorrect arrays.

Change-Id: I816989996fd5c7c46a84139dd7d871a081be354f