core: Add a freeable attribute to farray and fp16array.
This attribute determines whether or not the farray/fp16array object
contains a pointer to data buffer that was dynamically allocated and
therefore whether or not it can be freed. If not, a call to
free_farray() or free_fp16array() will only free the container struct,
not the data buffer.

It's becoming apparent that we need to share pointers in order to avoid
expensive memcpys. Without the language syntax support for smart pointer
types, this is a simple hack to get these optimizations to at least not
crash for us.

Change-Id: I0a3b263f35566a73c2020d53e1e8c702ef8b9108