all: In simulation, fast forward past all the initialization work.
This is most pronounced when loading large models, when FC weights
need to be compressed and packed, and/or when all values need to be
quantized to 16 bits. Eventually we'll be able to load in the data
directly in the desired format. For now, fast forwarding past this point
should reduce simulation times by quite a bit.

Change-Id: If35d136bd8f467fbe409dc632191d45456032831