smv: When skipping a convolution tile, mimic its memory behavior.
In the past, we would entirely skip tiles of a convolution when
sampling, which means we were also decreasing the workload's working set
size. To preserve the illusion of the same working set even for work
that gets skipped, we set the region of memory that would have been
modified by the accelerator to zero from the CPU.

Next, we will replace the CPU memset function calls with invocations of
a traffic generator to perform the memory accesses from an external
agent, therefore not polluting the L1 cache.

Change-Id: I1c84bcf7909187644a58be28bb017777cb894aea