smv: Add L2 cache aware tiling.
Our SMV's convolution dataflow will load all the kernels for every
input tile. This reloading of the weights is pretty expensive for
networks that have large kernels (like the large ELU network we are
using). To mitigate the reloading overhead, this L2 cache aware tiling
blocks a suitable number of kernels in an L2 tile, so that we shorten
the weights reuse distance, making better use of the L2 cache.

**CAVEATS**
The compromise of this L2-aware tiling is that now we need to reload
the entire input for every L2 tile, which introduces extra overhead.
Hence, use this smartly, otherwise it may lead to worse performance.

Change-Id: Ie1fcf6bbc1f9234291858f307cae021300f1faf7