all: Add debug ostream for debugging.
Instead of using std::cout and compile-time macros to disable
debugging output (such as the output values of each layer), use the
dout(int) function, which returns an ostream that either prints or
ignores the message based on the current debug level.

Debug levels are arbitrary integers. The global level is set with with
the --debug-level command line argument. If dout() is called with a
level between 0 and the global level, then it is printed; otherwise, it
is ignored.

Example:

  initDebugStream(2);
  dout(1) << "This will get printed" << std::endl;
  dout(2) << "This will get printed" << std::endl;
  dout(3) << "This will NOT get printed" << std::endl;

Change-Id: I54baef9970d744f627c54373b1baa60e1843790c