smv: Add convolution tiling optimizer.
The tiling optimizer examines a convolution layer shape and tiles the
inputs, weights, and outputs such that they can all fit in separate
scratchpad space. The optimizer first determines the dimensions along
which each tensor will be tiled, and then generates many possible tiling
configurations based on the selected dimensions. Finally, the optimizer
selects the configuration that maximizes the combined size of the input,
weights, and output tiles.

This commit implements the core of the optimizer: the part that
determines the basic tile shapes (which every tile will use, except
possibly the last tiles along a dimension). The next stage will actually
generate the tiled tensors and copy the data accordingly.

Change-Id: I2283026c61e25c6c6c1749c42ab7cc2586ad7306