core: Use factory methods to create operators.
This gives us the flexibility to define custom operators that are not
limited to explicit template specializations of the base operators.
The factory methods can return any type they want as long as it is
inherited from Operator, and we can rely on auto type deduction to keep
the code simple.

Change-Id: If58ebd1b5ad19efecbc7490aac8434c005d9f15e