modelconf: Generate random data invariant to data layout.
In the past, a randomly generated tensor in NCHW would physically look
the same but logically be different from one in NHWC. This makes the
logical output of each layer of the network dependent on layout, which
is obviously not desirable. Now, regardless of data layout, randomly
generated tensors will be logically consistent, in that they will always
represent the same "image" or "input", even if the underlying storage
order is different.

Change-Id: I217f52638aa267f92dfc19eded73f50355958bf4