python-client (3/5): Add C++ code to work with the Python client.
This mostly changes the read_model_conf.cpp, which now deserialize
a protobuf model file instead of a libconfuse file to build the
network. This also makes necessary change to tensor class in order
to directly create a Tensor from a TensorProto. Since the Python client
also takes care of data initilization, the data generator in C++ is
no longer used.

Change-Id: I71e7b8689493c1265e782006b0c745e6f92f0c2c