core: Remove the global backend.
This removes the reliance on the global backend in network_builder.
Instead, we use the backend information stored in the graph.

This has to temporarily drop the output printing in smaug.cpp, because
that also used the global backend. This functionality will be added
back in the next commit, which will remove the backend template in
the Tensor class, thus completely getting rid of the global backend.

Change-Id: I0bc4a224827dfe5052d6c482dc2fd8d142f2b1b5