core: Make tensors not backend templated.
The only backend-specific attribute in a tensor is the alignment (
in TensorShape). This removes the backend template from tensors. We
will need to specify the alignment explicitly (probably using a
factory method) when creating tensors.

Some global template functions become non-templated, so moved them
to the corresponding cpp files. Specifically, created a cpp file
for reorder_op_impl.h.

Also, in the Network class, getInput() now can't be overriden, since
the templated one becomes non-templated. To resolve that, I find that
we are not using the one that returns TensorBase, so I removed that.
The same for getOutput().

Somewhat surprisingly, all the referene tests passed.

Change-Id: I8b8d00c4dbb9f46f5a7cd1ed512ce97d0c62602c