python: Create DataOp for input tensors.
Previously we did this in network_builder. It's much easier to do
it in Python. And it will also be useful for the upcoming layouts
transformation commit.

Specifically, if any input tensor doesn't have a source operator,
we create a DataOp for it. Note that we don't need to create a DataOp
for input_data operator.

Change-Id: Id81f38fb2a6eba97cfdb4c277a5aa830cc1dcfc5