python: Check and add layout transformation if necessary.
This performs sort of automatic layout transformation at the Python
side. Layout transformation is required because different kernels
may expect different inputs/weights layouts. For example, SMV's
convolution operator requires both the input tensor and the filter
tensor in NHWC layout, whereas the user may have initialized these
two tensors in a differnt layout (i.e., NCHW).

Instead of resolving the layout differences before we actually
schedule the operator in the C++ core, I think it's more doable to add
reorder operators when we construct the graph in Python. Specifically,
when creating and adding operator to the graph, this will check
if any layout transformation is needed. And if yes, it adds reorder
operators to transform the tensors into expected layouts.

Bugfix: When we remove the extraneous reorder operators, we should not
alter the order in the parents/children fields of the node, otherwise
the network builder will find unexpected input order, which causes it
to set incorrect tensors in the operators. For example, for a convolution
operator, we expect the first input tensor to be the input activations and
the second input tensor to be the kernels.

Unit tests will be in the next commit.

Change-Id: I9be3b1fb2a568e89a2c3b4e72bb882324bb10c0d