smv: Fix a bug in setting TiledTensor shape.
So, I was stupid about the shape of a TiledTensor. When we set the
TensorShape for a TiledTensor, the alignemnt should always set to 0
(i.e., the default value). It's a shape that indicates where we tile
in the original tensor. No padding is needed. Otherwise, the
TensorIndexIterator would actually take the alignment into account,
causing unexpected errors.

This commit also adds the alignment to some TensorShapes in tests
where they do need it.

All tests passed except the one in the SMV convolution op tests, it
was marked as expected to fail because the convolution op impl was
WIP. Will look into that tomorrow.

Change-Id: I6a7f21eca03cc955975f0f66da0f6bd61793a870