core: Add a gem5 simulation option to the binary invocation.
This is towards using the same smaug binary for native run and gem5
simulation. To that end, this adds an option for gem5 simulation to
the smaug invocation. Then this removes the previous reliance on the
compile time constants to provide different operations in gem5/native
binaries. This also adds the necessary SSE flags for gem5 simulation.

P.S. The variadic template stuff actually took me several hours to
get around. But it works fine :)

Change-Id: Iec27dafff3f052668f60344dac46b673caa587f9