smv: Add an un-tile function for generating final output.
This adds a function for un-tiling a tiled tensor. This is useful for
generating the final output of an operator. After all the tiles
have been sent for computation, we get the results produced in a list
of tiled output tiles. We need to copy the results into the original
output tensor, so it can be used by subsequent operators.

Change-Id: Ifc4b78eb399e3058806a2a34ddde3e25874a221a