tests: Support output value verification for SMV convolution tests.
We manually compute and specify the expected output values in the
reference tests, which quickly becomes out of hand for operators with
large input/kernel sizes in the SMV tests.

This commit uses the reference convolution kernel for 'correct'
output values, as the simple reference implementation should
be pretty robust. A few reorder operators are needed to do the layout
transformations from NHWC to NCHW and vice versa.

This commit also adds a few tests. We will have more in the upcoming
commits.

Change-Id: I24a081b7bdbd2f337273743173aa7ac9b5ed4ba0