smv: Fix weight channelwise iteration.
When the weights are tiled channelwise, we need multiple kernel
invocations to finish all the weight channels. Since multiple weight
channelwise tiles produce the same output channel, we need to avoid
resetting the result for non-first weight tiles. Also, if the input
tile is not tiled channelwise, we need to provide the starting channel
of the input tile that will be effective for computation in the
invocation.

Change-Id: I65ea09fed2a515b3d1e0a57519c4b0a98c0b9cec