tests: Add more tests for SMV convolution kernel/tiling.
The added tests will cover more code paths such like:
1) Weight iteration for more than 8 kernels, both multiples of and
non-multiples of 8. This detects a bug when non-multiples of 8 kernels
is used and neither of the weights and the outputs need tiling. In this
case, the output tile generated by the tiling optimizer will contain
multiples of 8 kernels, which should be the original number of kernels.
2) More tests for DimNH tiling. As its interactions with same/valid
padding and input halos, we can easily forget something.
3) Weight channelwise iteration. When the input is not tiled channelwise,
the weight can be independently tiled channelwise.

TESTED=unit

Change-Id: I872436043547209fbd35da752d833464cad055e3