smv: Add DMA calls for the SMV convolution kernel.
We use float16 data for SMV operators, but because Aladdin doesn't
support float16 computation, we use float16 data for communication
and convert it to float32 for computation. This commit starts using
the local scratchpads for the SMV convolution kernel, and adds the
DMA calls for data loading/storing, which also handles data conversion.

This also changes the tests because:
1) The tests were using float32 data, and now all the SMV convolution
tests use float16 data. As a result, the scratchpads can now fit double
number of elements. Thus, a few tests need parameter adjustments in
order to serve their initial testing purposes.
2) The reference implementation still uses float32, to use it for
correct results, we now to need to convert SMV's float16 tensors into
using float32 for the reference operator. Likewise, the reference
results need float16 conversion before doing the validations.
3) To test the float16 and float32 coversion code, the values used for
tensor data initialization are modified to be float numbers. A bug in the
reference implementation is also detected by this.

I will add the array mapping calls in the next commit.

TESTED=unit

Change-Id: I29991ea18c10d1e4ea1de12c6fdefbcbb3f98daf