smv: Add output channelwise iteration when weights don't need N-wise tiling.
This commit is needed when the weights do not need N-wise tiling, whereas the
output needs to be tiled channelwise. This means the weight tile contains
more kernels that the output scratchpad can fit. Therefore, we need
multiple kernel invocations to finish the weight tile, where each invocation
only consumes part of it. To fix this, we need to provide the starting kernel
from which the weight tile will be effective for the invocation.

TESTED=unit

Change-Id: I82563abd44fe83b6e30ae7d294f732dcfb67709c