smv: Add DimNCH tiling.
This commit adds another tiling strategy which tiles over N, C and H
simultaneously. This covers most of the cases when none of DimN, DimNC and
DimNH could find a tiling solution. The size of this commit turns out to be
pretty small, thanks to the systematic design and modularity of the tiling
optimizer.

The DimNCH tiling is quite powerful. The added tests were all un-tilable in
the past. It wasn't very hard to run into an un-tilable layer with CIFAR10/100
(i.e., 32x32) inputs. Very nice to see that we are able to tile 128x128 images
now.

TESTED=unit

Change-Id: Ifd2973648b3860e023d687b5f8a3a06b85691276