tests: Add ELU network for testing.
This adds the ELU network for testing, which consists of 11 layers
of convolutions and 5 pooling layours. The two model files used in
the test contain the same parameters except for their target backends:
elu.pb runs on SMV while elu_ref.pb runs on Referemce. The size/number
of kernels used are identical to the original ELU network.

TESTED=unit

Change-Id: I128122b52565306bdeb87d345a6dbb85c741dccb