core: Use reorder operator for weights transposition.
This makes inner product weights transposition done ahead of the
operator creation. Specifically, it does the following things:
1) Add a new DataLayout called CN.
2) Move the 2D tensor transposition into the reorder operator. For a
reorder operator, if the source layout = NC (or CN) and the target
layout = CN (or NC), the 2D tensor will be transposed.
3) At the Python side, we were already automatically inserting reorder
ops for layout transformation. This commit augments this logic for
weights transposition. Basically, if the current weight layout is different
from the exepected one, a reorder op will be added.
4) Remove the weight transpositions in the SMV inner product tests.

This also helps the next commit that adds the reference transposed
inner product implementation. Testing is done together with that commit.

Change-Id: I482704e8e7a3db932cf6944597734354c06a17db