smv: Refactor the enumeration of tensor tiling configurations.
The SMV operators share a lot in common in enumeration of input
tensor tiling configs. This refactors out that part into two functions
in the base tiling optimizer. Each of them takes two vectors, one for
minimum dimensions of the tile and the other for stride sizes. To be
consistent, findBestTilingDims now also takes a vector for minimum
tile dimensions.

The refactored functions consider different tensor layouts because
some of the SMV operators may use different layouts, for example,
the batch norm operator will use NCHW.

Change-Id: Ie82dab1fa8da4829977db0e88317a09a1a0a218d