smv: Fix Rowwise tiling bug for convolution operator.
In DimNH (or DimNCH) tiling, due to stride size > 1, there is a case where
the last rowwise tile doesn't have enough rows for convolution. For example,
the last tile has 3 rows whereas the kernel size is 5x5. As a result, there
is no corresponding output tile for that input tile. Also, in the run()
function, we should not send this input tile for computation.

TESTED=unit

Change-Id: I232794b93f7b4fd9953713579ef91b0f76c1afce