python: Add one graph API to disable layout transformation.
This adds an API for the user to disable/enable the automic layout
transformation. As we are testing larger networks with the SMV backend,
this is useful when we want to create the same network with the reference
backend (the reference backend will by default tranforms NHWC to NCHW, but
it actually has NHWC implementations).

Change-Id: I7d0921d3a0a1d54ed71ef1d5fd970c393a9efe61