core: Use malloc_aligned and free to allocate/deallocate tensor data.
Because the DMA transfers are cacheline aligned, we also need to
cacheline align the tensor data. This fixes the crash in the ELU
network, of which the last layer has a 16-byte tile (not cacheline
aligned).

We need to zero out the tensor data because otherwise the padded area
will have uninitialized values.

TESTED=unit

Change-Id: I19c16da36ad921bdfefbc18867f1a93ec2cc1506