smv: Make tensor tiles have unique names.
A tensor can be outputs of an operator and at the same time be
inputs of other operators, if it needs to be tiled in different
operators, the current naming approach would generate tile
tensors with the same name. For example, tensor "t" is output of
operator "fc0" and input of operator "fc1". Both operators need
to tile tensor "t". We will end up having two tiles with the same name
"t/tile:0", where one is an input tile and the other is an output tile.
This leads to memory leaks (gem5 simulation crashes), because we rely
on the Workspace class to delete the tensors, which requires them to
have unique names.

To that end, because we already guarantee unique names for operators
(in the Python client), this commit concatenates the unique operator
name to that of the corresponding tile tensors.

TESTED=unit

Change-Id: Id7b86ce89165a0ce309343b81f4e089c8612b043