smv: Change an array name in the convolution kernel.
In Aladdin configuration, we don't yet have a way to differentiate
arrays with the same name in different top level function. Both the
convolution and the matrix multiply kernels have an array named
"product_reg", the former is 1024 bytes while the later is 128 bytes.
In order to avoid naming conflict, this renames the one in convolution
as "smv_conv_product_reg".

Change-Id: I42585f4b9c2c24dcb5a4db91e827b3c7bd8a081f