common: Fix Makefile for building tracer.
It didn't allow multi-threaded compilation to build tracer (really
slow), mostly because we generated the labelmap into one single file,
which needed to be serial. As we can now specify the labelmap output
file name to the LLVM-Tracer, both the labelmaps generation and the
LLVM objects building now become parallelizable.

Change-Id: Ifb93b5a7152d9f4ccfbb7ea50f45d9989eff2edd