operators: Fix rereading the inputs in some situations.
This commit has a cleaner way to decide whether the accelerator needs to read
the inputs/weights or not. Previously, we were actually unnecessarily reading
data in some cases. This commit keeps track of the last read input tile index,
and we will read a input tile only if its tile index is different from that of
last read tile.

TESTED=unit

Change-Id: I35079a92b11230aad5c33a21e7fa72bbfc9db8ac