operators: Change fp16 load and store functions into functions.
The fp16 load/store functions were inlined previously, but as we want to
characterize different host memory access mechanisms, it's better to remove
the inlining so that it creates a function call barrier, which isolates the
effect of overlap between data transfer and kernel computation. Otherwise, due
to the overlaps, it would be very difficult to get the execution time breakdown.
In SMAUGv1, we actually did something similar where we called the kernel
function as a normal function.

Change-Id: I1023ab5292259c39767f5fbc3c479d1126cb3f97