operators: Tile the operators before we run the network at all.
There are two parts involved in the tiling process. The first part is figuring
out the tile shape, and the second is filling data into the generated tiles.
The first part can be done before we run the network at all (i.e., offline),
since we already know the shape of the tensors. For the second part, we can
only pre-do that for weights, because they remain read-only whereas
inputs/outputs will change from layer to layer.

This commit tries to pre-do the tiling as much as possible.

TESTED=unit

Change-Id: Ia5bbbba4806f3e3f5fa28e313e8c147d81580e23