core: Don't copy data if a tensor doesn't need to be tiled.
Previously, even if a tensor doesn't need to be tiled (i.e., there's only one
tile in the generated TiledTensor), we will create a new tensor for the tile and
copy data to it from the original tensor. This commit eliminates this
unnecessary data copy by directly using the original tensor as the tile.

TESTED=unit

Change-Id: I0fef36e3c88e38f7aed78ba0f507a94e65ddc186