smv: Don't copy data for the output tiles.
The getTileWithData() API returns a tile with data copied from the original
tensor where the tile gets tiled from. For outputs, it's totally unnecessary to
copy data from the original tensor, as there's no useful data in it in the first
place. This commit eliminates the unnecessary data copying.

TESTED=unit

Change-Id: I41c446b14c903de60fe088d8040bb21dca03d05e