smv: Fix the rowwise output tile size in the convolution tiling optimizer.
The overall padding size is weightDim - 1, of which ceil(weightDim-1, 2) is
padded on the top, and the rest on the bottom (similarly for padding on left and
right). If rowwise tiling is used, we should use ceil(weightDim-1, 2) as the
padding size to compute the output tile size instead of weightDim-1. With this
fix, the tiling optimizer can potentially generate bigger input/output tiles,
which increases the scratchpad utilization, as well as fewer kernel invocations.

TESTED=unit

Change-Id: I656d8a1ec9829f41a06759983f2fb874a3482021