smv: Correct rowwise tiling for convolution op.
Previously, we used the generateTiledTensor() function to generate tiles for
input tiles. However, this would lead to inefficiency and incorrect results.
Take the first layer of ResNet50 as an example, the original input tensor shape
is (1, 225, 225, 3) and the weight tensor shape is (64, 7, 7, 3), with stride
size of 2x2. The tiling optimizer computes the input tile shape to be
(1, 9, 225, 3). And the call to generateTiledTensor() would generate 55 input
tiles of the same shape (1, 9, 225, 3). There are two problems with this rowwise
tiling:

1) Because generateTiledTensor doesn't account for input padding, the first tile
can have rows that will be discarded. The first tile has top padding of
ceil((7-1)/2)=3, so it has total rows of 3+9=12, with strides of 2x2, the tile
can do three rowwise convolutions (e.g., -3~3, -1~5 and 1~7), as shown below. We
can see that the last row of the tile is actually not used and will be
discarded.

2) This can also lead to incorrect results because the next tile starts from a
wrong place. Using the same example, the second tile starts from row 4 (halo
size is 7-2=5), so the second tile will do two rowwise convolutions (e.g., 4~10
and 6~12), where it should be 3~9 and 5~11.

First tile
___ ___
 p   ^
 p   |    ___
 p   |     ^
 0 conv0   |    ___
 1   |     |     ^
 2   v   conv1   |      Second tile
 3  ___    |     |      ___ ___
 4         v   conv2     4   ^
 5  _________    |       5   |    ___
 6               v       6   |     ^
 7  _______________      7 conv0   |
 8                       8   |     |
---                      9   v   conv1
                        10  ___    |
                        11         v
                        12  _________

The correct tiling (shown below) should have the first tile shape be
(1, 8, 225, 3) instead of (1, 9, 225, 3), so no rows are unused. And the second
tile starts from row 3.

To resolve these issues, this commit changes the generateTiledTensor function
with striding accounted for.

First tile
___ ___
 p   ^
 p   |    ___
 p   |     ^
 0 conv0   |    ___
 1   |     |     ^      Second tile
 2   v   conv1   |      ___ ___
 3  ___    |     |       3   ^
 4         v   conv2     4   |    ___
 5  _________    |       5   |     ^
 6               v       6 conv0   |
 7  _______________      7   |     |
___                      8   v   conv1
                         9  ___    |
                        10         v
                        11  _________

TESTED=unit

Change-Id: I2feb85650791f0ad3d93e7cbf9b4dfff4b66e161